test: Seed E2E tests data to make tests independant from signup

+++ api/e2etests/views.py
@@ -5,7 +5,9 @@ from rest_framework.authentication import TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from organisations.models import Subscription
+from environments.models import Environment
+from organisations.models import Organisation, OrganisationRole, Subscription
+from projects.models import Project
 from users.models import FFAdminUser
 
 from .permissions import E2ETestPermission
@@ -18,10 +20,29 @@ class Teardown(APIView):
 
     def post(self, request):
         # delete users created for e2e test by front end
-        if os.environ["FE_E2E_TEST_USER_EMAIL"]:
-            FFAdminUser.objects.filter(
-                email=os.environ["FE_E2E_TEST_USER_EMAIL"]
-            ).delete()
+        fe_e2e_test_user_email = os.environ["FE_E2E_TEST_USER_EMAIL"]
+        password = "pbkdf2_sha256$260000$RuWdd52KA8sgO7taxBtRH1$N81RojtN9tVUp9vd8nyBthp7EJZ+NYFrjn2QCm76QME="
+        if fe_e2e_test_user_email:
+            FFAdminUser.objects.filter(email=fe_e2e_test_user_email).delete()
+            FFAdminUser.objects.filter(email=fe_e2e_test_user_email + ".io").delete()
+            FFAdminUser.objects.filter(email="changeMail@test.com").delete()
+
+        # create user and organisation for e2e test by front end
+        organisation = Organisation.objects.create(name="Bullet Train Ltd")
+        org_admin = FFAdminUser.objects.create(
+            email=fe_e2e_test_user_email,
+            password=password,
+            username=fe_e2e_test_user_email,
+        )
+        org_admin.add_organisation(organisation, OrganisationRole.ADMIN)
+
+        project = Project.objects.create(
+            name="My Test Project", organisation=organisation
+        )
+
+        Environment.objects.create(name="Development", project=project)
+
+        Environment.objects.create(name="Production", project=project)
 
         return Response(status=status.HTTP_204_NO_CONTENT)
